feat(audit): introduce actor creation utilities and enhance booking audit context
- Introduced a new `impersonatedBy` field in the booking audit context to track users being impersonated, enhancing audit trail visibility.
- Updated relevant components and services to utilize the new actor utilities and context, ensuring consistent handling of impersonation data across the application.

This change improves the overall structure of actor management and enhances the auditing capabilities related to user impersonation.
