+++ packages/features/booking-audit/client/components/BookingHistory.tsx
@@ -43,6 +43,11 @@ type AuditLog = {
         displayEmail: string | null;
         displayAvatar: string | null;
     };
+    impersonatedBy?: {
+        displayName: string;
+        displayEmail: string | null;
+        displayAvatar: string | null;
+    } | null;
 };
 
 interface BookingLogsFiltersProps {
