+++ packages/features/ee/impersonation/lib/ImpersonationProvider.ts
@@ -96,9 +96,9 @@ const auditAndReturnNextUser = async (
     return {
       ...obj,
       impersonatedBy: {
-        id: impersonatedByUser?.id,
-        uuid: impersonatedByUser?.uuid,
-        role: impersonatedByUser?.role,
+        id: impersonatedByUser.id,
+        uuid: impersonatedByUser.uuid,
+        role: impersonatedByUser.role,
       },
     };
   }
